# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ultima Casa API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/skills": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "operationId": "6acb38b223d6471cc32eab45ea57be4b",
                "responses": {
                    "200": {
                        "description": "Get list of all skills"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Skills"
                ],
                "summary": "Create a new skill",
                "description": "Create a new skill with images",
                "operationId": "4725cd560d3046ae33bd690208e21d41",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SkillsCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Skill created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Skills"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/skills/{id}": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "summary": "Get skill details",
                "description": "Returns detailed information about a specific skill",
                "operationId": "81b9cbc27bbdd8b5a6591975bf75441a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Skill ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Skills"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Skill not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Skills"
                ],
                "summary": "Update an skill",
                "description": "Update skill information",
                "operationId": "6c932a7a6f40be346ab93e4dd28f14f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Skill ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SkillsCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skill updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Skills"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Skill not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Skills"
                ],
                "summary": "Delete an skill",
                "description": "Delete a specific skill",
                "operationId": "7401a82a4b0de61e80c67901e89d62dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Skill ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Skill deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Skill not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Skills": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "properties": {
                            "en": {
                                "type": "string"
                            },
                            "es": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SkillsCreate": {
                "properties": {
                    "name": {
                        "properties": {
                            "en": {
                                "type": "string"
                            },
                            "es": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}